Refactored jdbc
